+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [0.0.1+5]
 
 * Adds plural separator in I18NextOptions
+* Adds key separator in I18NextOptions
 
 ## [0.0.1+4]